Move the class into better location.
