+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Singers/Rows/Components/Timeline/SingerLyricTimeline.cs
@@ -14,7 +14,7 @@ using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Components.Containers;
 using osu.Game.Screens.Edit;
 
-namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Singers.Rows.Components
+namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Singers.Rows.Components.Timeline
 {
     [Cached]
     public partial class SingerLyricTimeline : BindableScrollContainer
